backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
firebase-admin>=5.0.3
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
python-dotenv~=1.0.1